Dedupe `hasFreeCouponTransfersOnly`
